Abort processing on usage of reserved context id
The current reserved context words are: site, content, page_store, global.
If any of these words are inserted into the context by a user script,
then rendering will fail. This does not apply to items added via the
[meta] key in frontmatter, as all items are prefixed with "meta" when
using this features.
