+++ cmslib/src/page.rs
@@ -165,75 +165,6 @@ impl LinkedAssets {
     }
 }
 
-#[derive(Debug)]
-pub struct RenderedPage {
-    pub html: String,
-    pub target: RetargetablePathBuf,
-}
-
-impl RenderedPage {
-    pub fn new<S: Into<String> + std::fmt::Debug>(html: S, target: &RetargetablePathBuf) -> Self {
-        Self {
-            html: html.into(),
-            target: target.clone(),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct RenderedPageCollection {
-    pages: Vec<RenderedPage>,
-}
-
-impl RenderedPageCollection {
-    pub fn from_iterable<T: Iterator<Item = RenderedPage>>(iterable: T) -> Self {
-        Self {
-            pages: iterable.collect::<Vec<_>>(),
-        }
-    }
-
-    pub fn from_vec(pages: Vec<RenderedPage>) -> Self {
-        Self { pages }
-    }
-
-    pub fn write_to_disk(&self) -> Result<(), std::io::Error> {
-        use std::fs;
-        for page in self.pages.iter() {
-            let target = page.target.to_path_buf();
-            util::make_parent_dirs(target.parent().expect("should have a parent path"))?;
-            let _ = fs::write(&target, &page.html)?;
-        }
-
-        Ok(())
-    }
-
-    pub fn iter<'a>(&'a self) -> std::slice::Iter<'a, RenderedPage> {
-        self.pages.iter()
-    }
-
-    pub fn as_slice(&self) -> &[RenderedPage] {
-        self.pages.as_slice()
-    }
-}
-
-impl IntoIterator for RenderedPageCollection {
-    type Item = RenderedPage;
-    type IntoIter = std::vec::IntoIter<Self::Item>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.pages.into_iter()
-    }
-}
-
-impl<'a> IntoIterator for &'a RenderedPageCollection {
-    type Item = &'a RenderedPage;
-    type IntoIter = std::slice::Iter<'a, RenderedPage>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.iter()
-    }
-}
-
 pub mod script {
     use rhai::plugin::*;
