+++ cmslib/src/render/mod.rs
@@ -1,6 +1,7 @@
 use std::path::Path;
 
 pub mod markup;
+pub mod page;
 pub mod template;
 
 #[derive(Debug)]
