+++ test/site-rules.rhai
@@ -43,4 +43,10 @@ rules.add_page_context("/index.md", |page| {
     })
   }
 });
+
+rules.add_page_context("/index.md", |page| {
+  new_context(#{
+    global: "crash"
+  })
+});
 rules
